Crate rusoto_cloud9

AWS Cloud9
AWS Cloud9 is a collection of tools that you can use to code, build, run, test, debug, and release software in the cloud.

For more information about AWS Cloud9, see the AWS Cloud9 User Guide.

AWS Cloud9 supports these operations:

CreateEnvironmentEC2: Creates an AWS Cloud9 development environment, launches an Amazon EC2 instance, and then connects from the instance to the environment.

CreateEnvironmentMembership: Adds an environment member to an environment.

DeleteEnvironment: Deletes an environment. If an Amazon EC2 instance is connected to the environment, also terminates the instance.

DeleteEnvironmentMembership: Deletes an environment member from an environment.

DescribeEnvironmentMemberships: Gets information about environment members for an environment.

DescribeEnvironments: Gets information about environments.

DescribeEnvironmentStatus: Gets status information for an environment.

ListEnvironments: Gets a list of environment identifiers.

UpdateEnvironment: Changes the settings of an existing environment.

UpdateEnvironmentMembership: Changes the settings of an existing environment member for an environment.

If you're using the service, you're probably looking for Cloud9Client and Cloud9.

Structs

Cloud9Client
A client for the AWS Cloud9 API.

CreateEnvironmentEC2Request

CreateEnvironmentEC2Result

CreateEnvironmentMembershipRequest

CreateEnvironmentMembershipResult

DeleteEnvironmentMembershipRequest

DeleteEnvironmentMembershipResult

DeleteEnvironmentRequest

DeleteEnvironmentResult

DescribeEnvironmentMembershipsRequest

DescribeEnvironmentMembershipsResult

DescribeEnvironmentStatusRequest

DescribeEnvironmentStatusResult

DescribeEnvironmentsRequest

DescribeEnvironmentsResult

Environment
Information about an AWS Cloud9 development environment.

EnvironmentLifecycle
Information about the current creation or deletion lifecycle state of an AWS Cloud9 development environment.

EnvironmentMember
Information about an environment member for an AWS Cloud9 development environment.

ListEnvironmentsRequest

ListEnvironmentsResult

UpdateEnvironmentMembershipRequest

UpdateEnvironmentMembershipResult

UpdateEnvironmentRequest

UpdateEnvironmentResult

Enums

CreateEnvironmentEC2Error
Errors returned by CreateEnvironmentEC2

CreateEnvironmentMembershipError
Errors returned by CreateEnvironmentMembership

DeleteEnvironmentError
Errors returned by DeleteEnvironment

DeleteEnvironmentMembershipError
Errors returned by DeleteEnvironmentMembership

DescribeEnvironmentMembershipsError
Errors returned by DescribeEnvironmentMemberships

DescribeEnvironmentStatusError
Errors returned by DescribeEnvironmentStatus

DescribeEnvironmentsError
Errors returned by DescribeEnvironments

ListEnvironmentsError
Errors returned by ListEnvironments

UpdateEnvironmentError
Errors returned by UpdateEnvironment

UpdateEnvironmentMembershipError
Errors returned by UpdateEnvironmentMembership

Traits

Cloud9
Trait representing the capabilities of the AWS Cloud9 API. AWS Cloud9 clients implement this trait.