Trait rusoto_cloud9::Cloud9

pub trait Cloud9 {
    fn create_environment_ec2(
        &self, 
        input: CreateEnvironmentEC2Request
    ) -> RusotoFuture<CreateEnvironmentEC2Result, CreateEnvironmentEC2Error>;

    fn create_environment_membership(
        &self, 
        input: CreateEnvironmentMembershipRequest
    ) -> RusotoFuture<CreateEnvironmentMembershipResult, CreateEnvironmentMembershipError>;

    fn delete_environment(
        &self, 
        input: DeleteEnvironmentRequest
    ) -> RusotoFuture<DeleteEnvironmentResult, DeleteEnvironmentError>;

    fn delete_environment_membership(
        &self, 
        input: DeleteEnvironmentMembershipRequest
    ) -> RusotoFuture<DeleteEnvironmentMembershipResult, DeleteEnvironmentMembershipError>;

    fn describe_environment_memberships(
        &self, 
        input: DescribeEnvironmentMembershipsRequest
    ) -> RusotoFuture<DescribeEnvironmentMembershipsResult, DescribeEnvironmentMembershipsError>;

    fn describe_environment_status(
        &self, 
        input: DescribeEnvironmentStatusRequest
    ) -> RusotoFuture<DescribeEnvironmentStatusResult, DescribeEnvironmentStatusError>;

    fn describe_environments(
        &self, 
        input: DescribeEnvironmentsRequest
    ) -> RusotoFuture<DescribeEnvironmentsResult, DescribeEnvironmentsError>;

    fn list_environments(
        &self, 
        input: ListEnvironmentsRequest
    ) -> RusotoFuture<ListEnvironmentsResult, ListEnvironmentsError>;

    fn update_environment(
        &self, 
        input: UpdateEnvironmentRequest
    ) -> RusotoFuture<UpdateEnvironmentResult, UpdateEnvironmentError>;

    fn update_environment_membership(
        &self, 
        input: UpdateEnvironmentMembershipRequest
    ) -> RusotoFuture<UpdateEnvironmentMembershipResult, UpdateEnvironmentMembershipError>;
}

Trait representing the capabilities of the AWS Cloud9 API. AWS Cloud9 clients implement this trait.

Required methods

fn create_environment_ec2(
&self,
input: CreateEnvironmentEC2Request
) -> RusotoFuture<CreateEnvironmentEC2Result, CreateEnvironmentEC2Error>

Creates an AWS Cloud9 development environment, launches an Amazon Elastic Compute Cloud (Amazon EC2) instance, and then connects from the instance to the environment.

fn create_environment_membership(
&self,
input: CreateEnvironmentMembershipRequest
) -> RusotoFuture<CreateEnvironmentMembershipResult, CreateEnvironmentMembershipError>

Adds an environment member to an AWS Cloud9 development environment.

fn delete_environment(
&self,
input: DeleteEnvironmentRequest
) -> RusotoFuture<DeleteEnvironmentResult, DeleteEnvironmentError>

Deletes an AWS Cloud9 development environment. If an Amazon EC2 instance is connected to the environment, also terminates the instance.

fn delete_environment_membership(
&self,
input: DeleteEnvironmentMembershipRequest
) -> RusotoFuture<DeleteEnvironmentMembershipResult, DeleteEnvironmentMembershipError>

Deletes an environment member from an AWS Cloud9 development environment.

fn describe_environment_memberships(
&self,
input: DescribeEnvironmentMembershipsRequest
) -> RusotoFuture<DescribeEnvironmentMembershipsResult, DescribeEnvironmentMembershipsError>

Gets information about environment members for an AWS Cloud9 development environment.

fn describe_environment_status(
&self,
input: DescribeEnvironmentStatusRequest
) -> RusotoFuture<DescribeEnvironmentStatusResult, DescribeEnvironmentStatusError>

Gets status information for an AWS Cloud9 development environment.

fn describe_environments(
&self,
input: DescribeEnvironmentsRequest
) -> RusotoFuture<DescribeEnvironmentsResult, DescribeEnvironmentsError>

Gets information about AWS Cloud9 development environments.

fn list_environments(
&self,
input: ListEnvironmentsRequest
) -> RusotoFuture<ListEnvironmentsResult, ListEnvironmentsError>

Gets a list of AWS Cloud9 development environment identifiers.

fn update_environment(
&self,
input: UpdateEnvironmentRequest
) -> RusotoFuture<UpdateEnvironmentResult, UpdateEnvironmentError>

Changes the settings of an existing AWS Cloud9 development environment.

fn update_environment_membership(
&self,
input: UpdateEnvironmentMembershipRequest
) -> RusotoFuture<UpdateEnvironmentMembershipResult, UpdateEnvironmentMembershipError>

Changes the settings of an existing environment member for an AWS Cloud9 development environment.

Loading content...